requirements_python313.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
diskcache==5.6.3
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.18.0
flake8==7.3.0
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
joblib==1.5.1
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
mccabe==0.7.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.1
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
